This is a bitch. Need to figure out and list what operations will have to be done in loops due to varying number of pleats etc. then create general subroutine for them.
Translating in PS works, so need to decide if I'm going to translate
in perl and keep coordinates the same in PS or use PS translations.

There is still the "canvas size" problem, but I thought maybe I can
do an initial PS translate to a point heuristically determined to fit
the whole bellows based on bellows dimensions. Need to see if PS has
a 'zoom to extants" magic